+++ src/app/cat-prep/tree/components/DetailsPanel.tsx
@@ -1,0 +1,38 @@
+"use client";
+
+import { Node } from "../models/node";
+import { Description } from "../models/description";
+
+// Right-side content panel
+export default function DetailsPanel({
+  selected,
+  descriptions,
+}: {
+  selected: Node | null;
+  descriptions: Description[];
+}) {
+  const desc = selected
+    ? descriptions.find((d) => d.parent_id === selected.id)
+    : null;
+
+  if (!selected) {
+    return (
+      <div className="max-w-2xl">
+        <h1 className="text-2xl font-bold mb-2">Welcome to CAT Preparation</h1>
+        <p className="text-gray-600">
+          Select any topic from the learning path on the left to view detailed explanations.
+        </p>
+      </div>
+    );
+  }
+
+  return (
+    <div className="max-w-2xl">
+      <h1 className="text-2xl font-bold mb-2">{selected.title}</h1>
+      <div className="text-sm text-gray-500 mb-4">{selected.type}</div>
+      <p className="leading-relaxed">
+        {desc?.text || "No description available for this topic yet."}
+      </p>
+    </div>
+  );
+}
