+++ src/app/cat-prep/tree/models/description.ts
@@ -1,0 +1,7 @@
+// Represents description text for a syllabus node
+export type Description = {
+  id: number;
+  parent_id: number;
+  text: string;
+  type: string;
+};
