+++ src/app/cat-prep/tree/models/node.ts
@@ -1,0 +1,9 @@
+// Represents a single syllabus node (subject / topic / subtopic)
+export type Node = {
+  id: number;
+  parent_id: number | null | "";
+  title: string;
+  type: string;
+  order_index: number;
+  children?: Node[];
+};
